// apps/api/src/db/reference-data/notification/notification.enum.ts
export enum NotificationType {
	PROGRAM_PREVISIONNEL = "PROGRAM_PREVISIONNEL",
	ORDER_STATUS_CHANGE = "ORDER_STATUS_CHANGE",
	ORDER_APPROVED = "ORDER_APPROVED",
	ORDER_REJECTED = "ORDER_REJECTED",
	ORDER_ASSIGNED = "ORDER_ASSIGNED",
	EXECUTION_COMPLETED = "EXECUTION_COMPLETED",
	DTM_RESPONSE = "DTM_RESPONSE",
	MATERIAL_AVAILABLE = "MATERIAL_AVAILABLE",
	CLAIM_CREATED = "CLAIM_CREATED",
	CLAIM_UPDATED = "CLAIM_UPDATED",
	ORDER_CREATED = "ORDER_CREATED",
	ORDER_MODIFIED = "ORDER_MODIFIED",
	SUB_ORDER_CREATED = "SUB_ORDER_CREATED",
}

export enum NotificationChannel {
	EMAIL = "EMAIL",
	SMS = "SMS",
	IN_APP = "IN_APP",
	PUSH = "PUSH",
}